Progress on organization
Button methods  moved to HandAndFootButtons.py,
HandManagement contains code that wraps the cards, rescales cards and readjusts XY locations and shows the hand.
