+++ source/client/HandView.py
@@ -4,19 +4,23 @@ import client.Button as Btn
 from client.ClickableImage import ClickableImage as ClickImg
 from client.CreateDisplay import CreateDisplay
 import client.HandAndFootButtons as HandAndFootButtons
+import client.HandManagement as HandManagement
 from client.UICardWrapper import UICardWrapper
-# todo: do I still need UICardWrapper here?  Be sure to check creating executable before removing it.
 import client.UIConstants as UIC
 from common.Card import Card
 from common.HandAndFoot import Deal_Size
 from common.HandAndFoot import Meld_Threshold
 
-class HandView:
-    """This class handles letting players actually input information
 
-    Drawing, melding and discards are all done in this class.
-    Player can also fidget with hand during other players' turns.
-    Parts of this version of HandView are customized for HandAndFoot.
+class HandView:
+    """This class handles player's cards and enables actions.
+
+    Actions are primarily performed using buttons, since these need to somewhat customized by game
+    the buttons are in HandAndFootButtons.py (it is Hand And Foot Specific).
+    Management of displaying the hand's cards is not game specific, and methods that help with that
+    are in HandManagement.py
+
+    Player can arrange their own hand, and prepare to play cards during other players' turns.
     """
     def __init__(self, controller, display):
         self.controller = controller
@@ -53,10 +57,8 @@ class HandView:
             self.labelMedium(str(self.pickup_pile_sz), 150, 35)
         HandAndFootButtons.CreateButtons(self)
 
-
-
     def update(self):
-        """This updates the view of the hand """
+        """This updates the view of the hand, between rounds it displays a message. """
 
         if (self.controller._state.round == -1):
             self.mesgBetweenRounds(self.betweenrounds)
@@ -79,10 +81,10 @@ class HandView:
         if len(self.current_hand) == 0:
             self.hand_info = []
         elif not self.last_hand == self.current_hand:
-            self.hand_info = self.wrapHand(self.current_hand, self.hand_info)
-        self.showHolding(self.hand_info)  # displays hand
+            self.hand_info = HandManagement.wrapHand(self, self.current_hand, self.hand_info)
+        HandManagement.showHolding(self, self.hand_info) # displays hand
         if self.refresh_flag:  # if needed to rescale card size, then refreshXY again.
-            self.hand_info = self.refreshXY(self.hand_info)
+            self.hand_info = HandManagement.refreshXY(self, self.hand_info)
         HandAndFootButtons.ButtonDisplay(self)
 
     def nextEvent(self):
@@ -116,7 +118,6 @@ class HandView:
                         elif element.status == 0:
                             element.status = 1
                             element.img_clickable.changeOutline(2)
-                # --- end of CardSelection  ---
 
             elif event.type == pygame.MOUSEMOTION:
                 HandAndFootButtons.MouseHiLight(self,pos)
@@ -134,9 +135,10 @@ class HandView:
                         if (color_index % 2) == 1:
                             color_index = color_index - 1
                             element.img_clickable.changeOutline(color_index)
-            # --- end of  Card_Highlighting  --
             # This next section has player enter desired values for wild cards.
             elif event.type == pygame.KEYDOWN and self.num_wilds > 0:
+                # todo:
+                # def assignWilds(self) -- tried to split this off and ran into a problem, and program kept crashing.
                 textnote = "Designate one of " + str(self.num_wilds) + "  wildcard(s)"
                 textnote = textnote + " enter value by typing:  1-9, 0 (for ten), j, q, k or a. "
                 acceptable_keys = self.wild_cards[0][1]
@@ -170,89 +172,10 @@ class HandView:
                         self.wild_cards = self.wild_cards[1:]
                     self.num_wilds = len(self.wild_cards)
 
-    def wrapHand(self, updated_hand, wrapped_hand):
-        """Associate each card in updated_hand with a UICardWrapper
 
-        if change is that new cards were added, then want to preserve location and status for cards that were
-        already in hand.
-        If cards were played, then the wrapped cards that are removed should to those cards with status = 2
-        (prepared cards).
-        Note that discards are removed immediately after controller confirms discard legal.
-        """
-        card_xy = (10, UIC.Table_Hand_Border + 40)
-        # sort cards so that if prepared cards were played, those are the instances of the cards that are removed.
-        old_wrapped_hand = sorted(wrapped_hand, key = lambda x: x.status)
-        updated_wrapped_hand = []
-        if not updated_hand == []:
-            for card in updated_hand:
-                newcard = True
-                for already_wrapped in old_wrapped_hand:
-                    if newcard and card == already_wrapped.card:
-                        card_wrapped = already_wrapped
-                        card_xy = (max(card_xy[0], card_wrapped.img_clickable.x), card_xy[1])
-                        old_wrapped_hand.remove(already_wrapped)
-                        newcard = False
-                if newcard:
-                    card_xy = (card_xy[0] + self.hand_scaling[1], card_xy[1])
-                    card_wrapped = UICardWrapper(card, card_xy, self.hand_scaling[0])
-                updated_wrapped_hand.append(card_wrapped)
-            maxX = UIC.Disp_Width - (self.hand_scaling[1] / 2)
-            if (card_xy[0] > maxX):
-                scalingfactor = maxX / card_xy[0]
-                self.hand_scaling = (scalingfactor * self.hand_scaling[0], scalingfactor * self.hand_scaling[1])
-                updated_wrapped_hand = self.rescaleCards(updated_wrapped_hand, self.hand_scaling[0])
-                self.refresh_flag = True
-            if (self.hand_scaling[0] != UIC.scale and len(updated_wrapped_hand) <= self.deal_size):
-                self.hand_scaling = (UIC.scale, UIC.Card_Spacing)
-                updated_wrapped_hand = self.rescaleCards(updated_wrapped_hand, self.hand_scaling[0])
-            return updated_wrapped_hand
 
-    def refreshXY(self, original, layout_option=1):
-        self.refresh_flag = False
-        """After sorting or melding, may wish to refresh card's xy coordinates """
-        maxX = UIC.Disp_Width - (self.hand_scaling[1] / 2)
-        if not layout_option == 1:
-            print('the only layout supported now is cards in a line, left to right')
-        refreshed = []
-        card_xy = (10, UIC.Table_Hand_Border + 40)
-        for element in original:
-            element.img_clickable.x = card_xy[0]
-            element.img_clickable.y = card_xy[1]
-            card_xy = (card_xy[0] + self.hand_scaling[1], card_xy[1])
-            refreshed.append(element)
-        if (card_xy[0] > maxX):
-            scalingfactor = maxX / card_xy[0]
-            self.hand_scaling = (scalingfactor * self.hand_scaling[0], scalingfactor * self.hand_scaling[1])
-            refreshed = self.rescaleCards(refreshed, self.hand_scaling[0])
-        return refreshed
-
-    def rescaleCards(self, original, card_scaling):
-        rescaled = []
-        for element in original:
-            loc_xy = (element.img_clickable.x, element.img_clickable.y)
-            scaledelement = UICardWrapper(element.card, loc_xy, card_scaling)
-            scaledelement.status = element.status
-            scaledelement.img_clickable.outline_index = element.img_clickable.outline_index
-            rescaled.append(scaledelement)
-        self.refresh_flag = True
-        return rescaled
-
-    def showHolding(self, wrapped_cards):
-        wrapped_cards.sort(key=lambda wc: wc.img_clickable.x)
-        for wrapped_element in wrapped_cards:
-            color = UIC.outline_colors[wrapped_element.img_clickable.outline_index]
-            loc_xy = (wrapped_element.img_clickable.x, wrapped_element.img_clickable.y)
-            wrapped_element.img_clickable.draw(self.display, loc_xy, color)
-
-    def gatherSelected(self):
-        self.selected_list = []
-        for element in self.hand_info:
-            if element.status == 1:
-                self.selected_list.append(element)
-        return self.selected_list
-
-    # Confirm a user is sure about a discard and then perform it once confirmed.
     def discardConfirmation(self, confirmed, wrapped_discards):
+        # Confirm a user is sure about a discard and then perform it once confirmed.
         discards = []
         for element in wrapped_discards:
             discards.append(element.card)
@@ -270,11 +193,7 @@ class HandView:
                     for element in wrapped_discards:
                         self.hand_info.remove(element)
             return False # now that this is done, we don't have anything waiting on confirmation
-    def clearPreparedCardsGui(self):
-        for element in self.hand_info:
-            if element.status == 2:
-                element.status = 0
-                element.img_clickable.changeOutline(0)
+
     def mesgBetweenRounds(self, results):
         # print results where cards usually go until Ready button is clicked for next round.
         font = UIC.Medium_Text
